rename from operations to automations
