+++ ingestion/tests/unit/test_workflow_parse.py
@@ -16,7 +16,7 @@ from unittest import TestCase
 
 from pydantic import ValidationError
 
-from metadata.generated.schema.entity.operations.testServiceConnection import (
+from metadata.generated.schema.entity.automations.testServiceConnection import (
     TestServiceConnectionRequest,
 )
 from metadata.generated.schema.entity.services.connections.dashboard.tableauConnection import (
