+++ openmetadata-airflow-apis/openmetadata_managed_apis/operations/test_connection.py
@@ -21,7 +21,7 @@ from openmetadata_managed_apis.workflows.ingestion.credentials_builder import (
     build_secrets_manager_credentials,
 )
 
-from metadata.generated.schema.entity.operations.testServiceConnection import (
+from metadata.generated.schema.entity.automations.testServiceConnection import (
     TestServiceConnectionRequest,
 )
 from metadata.ingestion.connections.test_connections import SourceConnectionException
