+++ openmetadata-service/src/main/java/org/openmetadata/service/clients/pipeline/noop/NoopClient.java
@@ -18,8 +18,8 @@ import java.util.Map;
 import javax.ws.rs.core.Response;
 import org.openmetadata.schema.ServiceEntityInterface;
 import org.openmetadata.schema.api.configuration.pipelineServiceClient.PipelineServiceClientConfiguration;
-import org.openmetadata.schema.entity.operations.TestServiceConnectionRequest;
-import org.openmetadata.schema.entity.operations.Workflow;
+import org.openmetadata.schema.entity.automations.TestServiceConnectionRequest;
+import org.openmetadata.schema.entity.automations.Workflow;
 import org.openmetadata.schema.entity.services.ingestionPipelines.IngestionPipeline;
 import org.openmetadata.schema.entity.services.ingestionPipelines.PipelineStatus;
 import org.openmetadata.sdk.PipelineServiceClient;
@@ -44,7 +44,7 @@ public class NoopClient extends PipelineServiceClient {
   }
 
   @Override
-  public Response runOperationsWorkflow(Workflow workflow) {
+  public Response runAutomationsWorkflow(Workflow workflow) {
     return null;
   }
