+++ openmetadata-service/src/main/java/org/openmetadata/service/jdbi3/CollectionDAO.java
@@ -60,6 +60,7 @@ import org.openmetadata.schema.entity.Bot;
 import org.openmetadata.schema.entity.Type;
 import org.openmetadata.schema.entity.alerts.Alert;
 import org.openmetadata.schema.entity.alerts.AlertAction;
+import org.openmetadata.schema.entity.automations.Workflow;
 import org.openmetadata.schema.entity.classification.Classification;
 import org.openmetadata.schema.entity.classification.Tag;
 import org.openmetadata.schema.entity.data.Chart;
@@ -76,7 +77,6 @@ import org.openmetadata.schema.entity.data.Pipeline;
 import org.openmetadata.schema.entity.data.Report;
 import org.openmetadata.schema.entity.data.Table;
 import org.openmetadata.schema.entity.data.Topic;
-import org.openmetadata.schema.entity.operations.Workflow;
 import org.openmetadata.schema.entity.policies.Policy;
 import org.openmetadata.schema.entity.services.DashboardService;
 import org.openmetadata.schema.entity.services.DatabaseService;
@@ -3450,7 +3450,7 @@ public interface CollectionDAO {
   interface WorkflowDAO extends EntityDAO<Workflow> {
     @Override
     default String getTableName() {
-      return "operations_workflow";
+      return "automations_workflow";
     }
 
     @Override
