+++ openmetadata-service/src/main/java/org/openmetadata/service/resources/operations/WorkflowResource.java
@@ -36,10 +36,10 @@ import javax.ws.rs.core.SecurityContext;
 import javax.ws.rs.core.UriInfo;
 import lombok.extern.slf4j.Slf4j;
 import org.openmetadata.schema.api.data.RestoreEntity;
-import org.openmetadata.schema.entity.operations.CreateWorkflow;
-import org.openmetadata.schema.entity.operations.Workflow;
-import org.openmetadata.schema.entity.operations.WorkflowStatus;
-import org.openmetadata.schema.entity.operations.WorkflowType;
+import org.openmetadata.schema.entity.automations.CreateWorkflow;
+import org.openmetadata.schema.entity.automations.Workflow;
+import org.openmetadata.schema.entity.automations.WorkflowStatus;
+import org.openmetadata.schema.entity.automations.WorkflowType;
 import org.openmetadata.schema.services.connections.metadata.OpenMetadataConnection;
 import org.openmetadata.schema.type.EntityHistory;
 import org.openmetadata.schema.type.Include;
@@ -104,7 +104,7 @@ public class WorkflowResource extends EntityResource<Workflow, WorkflowRepositor
   @Operation(
       operationId = "listWorkflows",
       summary = "List operations workflows",
-      tags = "operationsWorkflow",
+      tags = "automationsWorkflow",
       description =
           "Get a list of operations workflows. Use `fields` "
               + "parameter to get only necessary fields. Use cursor-based pagination to limit the number "
@@ -170,7 +170,7 @@ public class WorkflowResource extends EntityResource<Workflow, WorkflowRepositor
   @Operation(
       operationId = "listAllWorkflowVersion",
       summary = "List Workflow versions",
-      tags = "operationsWorkflow",
+      tags = "automationsWorkflow",
       description = "Get a list of all the versions of a Workflow identified by `Id`",
       responses = {
         @ApiResponse(
@@ -190,7 +190,7 @@ public class WorkflowResource extends EntityResource<Workflow, WorkflowRepositor
   @Path("/{id}")
   @Operation(
       summary = "Get a Workflow by Id",
-      tags = "operationsWorkflow",
+      tags = "automationsWorkflow",
       description = "Get a Workflow by `Id`.",
       responses = {
         @ApiResponse(
@@ -223,7 +223,7 @@ public class WorkflowResource extends EntityResource<Workflow, WorkflowRepositor
   @Operation(
       operationId = "getWorkflowByName",
       summary = "Get a Workflow by name",
-      tags = "operationsWorkflow",
+      tags = "automationsWorkflow",
       description = "Get a Workflow by `name`.",
       responses = {
         @ApiResponse(
@@ -257,7 +257,7 @@ public class WorkflowResource extends EntityResource<Workflow, WorkflowRepositor
   @Operation(
       operationId = "getSpecificWorkflowVersion",
       summary = "Get a version of the Workflow",
-      tags = "operationsWorkflow",
+      tags = "automationsWorkflow",
       description = "Get a version of the Workflow by given `Id`",
       responses = {
         @ApiResponse(
@@ -285,7 +285,7 @@ public class WorkflowResource extends EntityResource<Workflow, WorkflowRepositor
   @Operation(
       operationId = "createWorkflow",
       summary = "Create a Workflow",
-      tags = "operationsWorkflow",
+      tags = "automationsWorkflow",
       description = "Create a Workflow.",
       responses = {
         @ApiResponse(
@@ -307,7 +307,7 @@ public class WorkflowResource extends EntityResource<Workflow, WorkflowRepositor
   @Operation(
       operationId = "triggerWorkflow",
       summary = "Trigger an workflow run",
-      tags = "operationsWorkflow",
+      tags = "automationsWorkflow",
       description = "Trigger a workflow run by id.",
       responses = {
         @ApiResponse(
@@ -316,14 +316,14 @@ public class WorkflowResource extends EntityResource<Workflow, WorkflowRepositor
             content = @Content(mediaType = "application/json")),
         @ApiResponse(responseCode = "404", description = "Workflow for instance {id} is not found")
       })
-  public Response runOperationsWorkflow(
+  public Response runAutomationsWorkflow(
       @Context UriInfo uriInfo,
       @Parameter(description = "Id of the Workflow", schema = @Schema(type = "UUID")) @PathParam("id") UUID id,
       @Context SecurityContext securityContext)
       throws IOException {
     EntityUtil.Fields fields = getFields(FIELD_OWNER);
     Workflow workflow = dao.get(uriInfo, id, fields);
-    return pipelineServiceClient.runOperationsWorkflow(workflow);
+    return pipelineServiceClient.runAutomationsWorkflow(workflow);
   }
 
   @PATCH
@@ -331,7 +331,7 @@ public class WorkflowResource extends EntityResource<Workflow, WorkflowRepositor
   @Operation(
       operationId = "patchWorkflow",
       summary = "Update a Workflow",
-      tags = "operationsWorkflow",
+      tags = "automationsWorkflow",
       description = "Update an existing Workflow using JsonPatch.",
       externalDocs = @ExternalDocumentation(description = "JsonPatch RFC", url = "https://tools.ietf.org/html/rfc6902"))
   @Consumes(MediaType.APPLICATION_JSON_PATCH_JSON)
@@ -356,7 +356,7 @@ public class WorkflowResource extends EntityResource<Workflow, WorkflowRepositor
   @Operation(
       operationId = "createOrUpdateWorkflow",
       summary = "Update Workflow",
-      tags = "operationsWorkflow",
+      tags = "automationsWorkflow",
       description = "Create a Workflow, if it does not exist, or update an existing Workflow.",
       responses = {
         @ApiResponse(
@@ -376,7 +376,7 @@ public class WorkflowResource extends EntityResource<Workflow, WorkflowRepositor
   @Operation(
       operationId = "deleteWorkflow",
       summary = "Delete a Workflow",
-      tags = "operationsWorkflow",
+      tags = "automationsWorkflow",
       description = "Delete a Workflow by `id`.",
       responses = {
         @ApiResponse(responseCode = "200", description = "OK"),
@@ -399,7 +399,7 @@ public class WorkflowResource extends EntityResource<Workflow, WorkflowRepositor
   @Operation(
       operationId = "deleteWorkflowByName",
       summary = "Delete a Workflow",
-      tags = "operationsWorkflow",
+      tags = "automationsWorkflow",
       description = "Delete a Workflow by `name`.",
       responses = {
         @ApiResponse(responseCode = "200", description = "OK"),
@@ -423,7 +423,7 @@ public class WorkflowResource extends EntityResource<Workflow, WorkflowRepositor
   @Operation(
       operationId = "restore",
       summary = "Restore a soft deleted Workflow",
-      tags = "operationsWorkflow",
+      tags = "automationsWorkflow",
       description = "Restore a soft deleted Workflow.",
       responses = {
         @ApiResponse(
