+++ openmetadata-service/src/test/java/org/openmetadata/service/pipelineService/MockPipelineServiceClient.java
@@ -5,8 +5,8 @@ import java.util.Map;
 import javax.ws.rs.core.Response;
 import org.openmetadata.schema.ServiceEntityInterface;
 import org.openmetadata.schema.api.configuration.pipelineServiceClient.PipelineServiceClientConfiguration;
-import org.openmetadata.schema.entity.operations.TestServiceConnectionRequest;
-import org.openmetadata.schema.entity.operations.Workflow;
+import org.openmetadata.schema.entity.automations.TestServiceConnectionRequest;
+import org.openmetadata.schema.entity.automations.Workflow;
 import org.openmetadata.schema.entity.services.ingestionPipelines.IngestionPipeline;
 import org.openmetadata.schema.entity.services.ingestionPipelines.PipelineStatus;
 import org.openmetadata.sdk.PipelineServiceClient;
@@ -28,7 +28,7 @@ public class MockPipelineServiceClient extends PipelineServiceClient {
   }
 
   @Override
-  public Response runOperationsWorkflow(Workflow workflow) {
+  public Response runAutomationsWorkflow(Workflow workflow) {
     return null;
   }
