+++ openmetadata-service/src/test/java/org/openmetadata/service/resources/operations/WorkflowResourceTest.java
@@ -9,10 +9,10 @@ import static org.openmetadata.service.util.TestUtils.assertListNull;
 import java.io.IOException;
 import java.util.Map;
 import org.apache.http.client.HttpResponseException;
-import org.openmetadata.schema.entity.operations.CreateWorkflow;
-import org.openmetadata.schema.entity.operations.TestServiceConnectionRequest;
-import org.openmetadata.schema.entity.operations.Workflow;
-import org.openmetadata.schema.entity.operations.WorkflowType;
+import org.openmetadata.schema.entity.automations.TestServiceConnectionRequest;
+import org.openmetadata.schema.entity.automations.Workflow;
+import org.openmetadata.schema.entity.automations.WorkflowType;
+import org.openmetadata.schema.entity.automations.CreateWorkflow;
 import org.openmetadata.schema.entity.services.ServiceType;
 import org.openmetadata.schema.services.connections.database.MysqlConnection;
 import org.openmetadata.service.Entity;
